requirements.txt
addict==2.4.0
anyio==4.6.2.post1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
breezy==3.2.1
Brlapi==0.8.3
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
debugpy==1.8.9
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
dulwich==0.20.31
duplicity==0.8.21
empy==3.3.4
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
fastjsonschema==2.20.0
filelock==3.16.1
filterpy==1.4.5
Flask==3.0.3
fonttools==4.55.0
fqdn==1.5.1
fsspec==2024.10.0
future==0.18.2
genmsg==0.5.16
genpy==0.6.16
gpg==1.16.0
gymnasium==1.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.27.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipympl==0.9.4
ipython==8.29.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==23.5.0
kiwisolver==1.4.7
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.3.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
meld==3.20.4
mercurial==6.1.1
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpngw==0.1.4
numpy==1.26.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
open3d==0.18.0
overrides==7.7.0
packaging==24.2
panda-gym==3.0.7
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
patiencediff==0.2.1
pexpect==4.8.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==3.12.4
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybullet==3.2.6
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyquaternion==0.9.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.9.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robotic==0.2.1
roman==3.3
rosclean==1.15.8
rosdep==0.21.0
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
rosmaster==1.15.14
rosparam==1.15.14
rospkg==1.6.0
rospkg-modules==1.6.0
rpds-py==0.21.0
scikit-learn==1.6.0
scipy==1.14.1
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
systemd-python==234
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.1.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.13.3
wstool==0.1.18
xdg==5
xkit==0.0.0
zipp==1.0.0
